Rename package and fix grammar
