+++ chain/d3nlib/utils.go
@@ -1,4 +1,4 @@
-package provider
+package d3nlib
 
 import (
 	"os"
